SwiftUI facade: @BlazeQuery as default, env injection, slim getting-started doc
- Add typed @BlazeQuery (BlazeDocument), env-based blazeDBClient, BlazeDataQuery for raw rows
- BlazeQueryTyped typealias + deprecated type: inits; optional db + primitive equals overloads
- BlazeQueryTypedObserver late bind for env-only path; tests for lazy bind + alias
- Docs: blessed path in README/SWIFTUI_INTEGRATION; migration note; front-door SWIFTUI_DATABASE_PATTERNS
- Examples and tests updated for BlazeDataQuery rename

Made-with: Cursor
