+++ Docs/GettingStarted/SWIFTUI_DATABASE_PATTERNS.md
@@ -1,13 +1,18 @@
 # BlazeDB in SwiftUI
 
-BlazeDB is not tied to SwiftUI. This guide shows one way to use it in a SwiftUI app.
+**Standard SwiftUI pattern**
 
-## Level 1 - Simple (start here)
+- Open BlazeDB **once**
+- Inject the client **once** at the app root
+- Read typed models with **`@BlazeQuery`**
+- Write through **`@Environment(\.blazeDBClient)`**
+- Add a **store** only when write logic grows
 
-Use this when: you want the fastest path to a working app.
+For deeper reference (explicit `db:`, raw queries, compatibility, custom environment), see the [SwiftUI Integration Guide](../Guides/SWIFTUI_INTEGRATION.md). Upgrading old examples: [SwiftUI facade migration](SWIFTUI_FACADE_MIGRATION.md).
 
-Open BlazeDB once, keep it in one app object, and pass it into your screens.
-Think of `AppDatabase` as the place your app keeps its database.
+---
+
+## Level 1 — Standard pattern
 
 ```swift
 import SwiftUI
@@ -15,165 +20,138 @@ import BlazeDB
 
 final class AppDatabase {
     static let shared = AppDatabase()
-    let db = try! BlazeDB.open(name: "myapp", password: "Password123!")
+    let db: BlazeDBClient
+
+    private init() {
+        self.db = try! BlazeDB.open(
+            name: "myapp",
+            password: "Password123!"
+        )
+    }
+}
+
+struct TodoItem: BlazeDocument {
+    var id: UUID
+    var title: String
+    var isDone: Bool
+
+    init(id: UUID = UUID(), title: String, isDone: Bool = false) {
+        self.id = id
+        self.title = title
+        self.isDone = isDone
+    }
+
+    init(from storage: BlazeDataRecord) throws {
+        guard let id = storage.storage["id"]?.uuidValue,
+              let title = storage.storage["title"]?.stringValue else {
+            throw BlazeDBError.invalidData(reason: "Invalid TodoItem")
+        }
+        self.id = id
+        self.title = title
+        self.isDone = storage.storage["isDone"]?.boolValue ?? false
+    }
+
+    func toStorage() throws -> BlazeDataRecord {
+        BlazeDataRecord([
+            "id": .uuid(id),
+            "title": .string(title),
+            "isDone": .bool(isDone)
+        ])
+    }
 }
 
 @main
 struct MyApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView(database: AppDatabase.shared)
+            ContentView()
+                .blazeDBEnvironment(AppDatabase.shared.db)
         }
     }
 }
 
-struct TodoItem: BlazeStorable {
-    var id: UUID = UUID()
-    var title: String
-    var isDone: Bool = false
-}
-
 struct ContentView: View {
-    let database: AppDatabase
-    @State private var items: [TodoItem] = []
+    @Environment(\.blazeDBClient) private var database
+    @BlazeQuery var items: [TodoItem]
 
     var body: some View {
         VStack {
             Button("Add Sample Item") {
-                do {
-                    try database.db.put(TodoItem(title: "Buy milk"))
-                    items = try database.db.query("todoitem").all()
-                } catch {
-                    print(error)
-                }
+                try? database?.insert(TodoItem(title: "Buy milk"))
             }
-
             List(items, id: \.id) { item in
                 Text(item.title)
             }
         }
-        .task {
-            do {
-                items = try database.db.query("todoitem").all()
-            } catch {
-                print(error)
+    }
+}
+```
+
+**Why this is the standard**
+
+- BlazeDB opens once for the app.
+- `@BlazeQuery` keeps the list in sync with the database.
+- Views do not need custom query wiring or passing `BlazeDBClient` through every initializer.
+
+---
+
+## Level 2 — Add a store when writes grow
+
+Keep reads in the view with `@BlazeQuery`. Move validation and multi-step writes into a store when the screen gets heavier.
+
+```swift
+@MainActor
+final class TodoWriteStore: ObservableObject {
+    func addSample(using database: BlazeDBClient?) {
+        try? database?.insert(TodoItem(title: "From store"))
+    }
+}
+
+struct ContentView: View {
+    @Environment(\.blazeDBClient) private var database
+    @StateObject private var store = TodoWriteStore()
+    @BlazeQuery var items: [TodoItem]
+
+    var body: some View {
+        VStack {
+            Button("Add via store") {
+                store.addSample(using: database)
+            }
+            List(items, id: \.id) { item in
+                Text(item.title)
             }
         }
     }
 }
 ```
 
-Avoid this:
-- Do not open BlazeDB inside `@State`.
-- Do not open BlazeDB inside a view `body`.
+---
+
+## Level 3 — Larger apps
+
+- One shared `BlazeDBClient` (or `AppDatabase`) for the process.
+- Use `@BlazeQuery` in feature views.
+- Add a store per feature when that feature needs coordinated writes.
 
 ---
 
-## Level 2 - Use a store (cleaner UI code)
+## Advanced patterns
 
-Use this when: one screen starts having more buttons, states, and DB calls.
+Use the [SwiftUI Integration Guide](../Guides/SWIFTUI_INTEGRATION.md) for:
 
-Move read/write logic into a store so the view mostly handles UI.
-
-```swift
-import Foundation
-import BlazeDB
-
-final class TodoStore: ObservableObject {
-    private let database: AppDatabase
-    @Published var items: [TodoItem] = []
-
-    init(database: AppDatabase) {
-        self.database = database
-    }
-
-    func load() {
-        items = (try? database.db.query("todoitem").all()) ?? []
-    }
-
-    func add(_ title: String) {
-        do {
-            try database.db.put(TodoItem(title: title))
-            load()
-        } catch {
-            print(error)
-        }
-    }
-}
-```
-
-Use it in a screen:
-
-```swift
-struct ContentView: View {
-    @StateObject private var store = TodoStore(database: AppDatabase.shared)
-
-    var body: some View {
-        List(store.items, id: \.id) { item in
-            Text(item.title)
-        }
-        .task {
-            store.load()
-        }
-    }
-}
-```
+- explicit `db:` on query wrappers  
+- raw `@BlazeDataQuery`  
+- compatibility aliases and edge cases  
+- custom environment wrappers  
 
 ---
 
-## Level 3 - Larger apps (multiple features)
+## Summary
 
-Use this when: your app has separate domains (for example Notes and Tasks).
+Default SwiftUI path:
 
-Each store handles its own feature, but they all use the same database.
+- `.blazeDBEnvironment(AppDatabase.shared.db)`
+- `@BlazeQuery` for typed lists  
+- `@Environment(\.blazeDBClient)` for writes (`insert` for `BlazeDocument` models)
 
-```swift
-struct Note: BlazeStorable {
-    var id: UUID = UUID()
-    var title: String
-}
-
-final class NotesStore: ObservableObject {
-    private let database: AppDatabase
-    @Published var notes: [Note] = []
-    init(database: AppDatabase) { self.database = database }
-
-    func load() {
-        notes = (try? database.db.query("note").all()) ?? []
-    }
-}
-
-final class TasksStore: ObservableObject {
-    private let database: AppDatabase
-    init(database: AppDatabase) { self.database = database }
-}
-```
-
----
-
-## Level 4 - Advanced (optional patterns)
-
-Use this when: you want cleaner dependency wiring for bigger apps and testing.
-This is optional. You do not need this for most apps.
-This is useful for large apps or testing, but most apps do not need it.
-
-```swift
-import SwiftUI
-
-struct DatabaseKey: EnvironmentKey {
-    static let defaultValue: AppDatabase = .shared
-}
-
-extension EnvironmentValues {
-    var database: AppDatabase {
-        get { self[DatabaseKey.self] }
-        set { self[DatabaseKey.self] = newValue }
-    }
-}
-```
-
-Then in a view:
-
-```swift
-@Environment(\.database) private var database
-```
+That is the front door. Details live in the integration guide and migration doc.
